pyproject.toml
[tool.poetry]
name = "fast_api_postgres_enterprise"
version = "0.1.0"
description = ""
authors = ["Michael Asumadu <masumadu@archsolutions.online>"]

[tool.poetry.dependencies]
python = "^3.9"
pytz = "2022.7"
alembic = "^1.9.1"
anyio = "^3.6.2"
app = "^0.0.1"
attrs = "^22.2.0"
certifi = "^2022.12.7"
charset-normalizer = "3.0.1"
click = "8.1.3"
dnspython = "2.2.1"
email-validator = "1.3.0"
fastapi = "0.88.0"
fastapi-utils = "0.2.1"
greenlet = "2.0.1"
h11 = "0.14.0"
httpcore = "0.16.3"
httptools = "0.5.0"
httpx = "0.23.1"
idna = "3.4"
importlib-metadata = "6.0.0"
itsdangerous = "2.1.2"
Jinja2 = "3.1.2"
Mako = "1.2.4"
MarkupSafe = "2.1.1"
more-itertools = "9.0.0"
orjson = "3.8.3"
packaging = "22.0"
pip = "22.3.1"
pluggy = "0.13.1"
psycopg2-binary = "2.9.5"
py = "1.11.0"
pydantic = "1.10.2"
PyJWT = "2.6.0"
python-dateutil = "2.8.2"
python-dotenv = "0.21.0"
python-multipart = "0.0.5"
PyYAML = "6.0"
requests = "2.28.2"
rfc3986 = "1.5.0"
setuptools = "63.1.0"
six = "1.16.0"
sniffio = "1.3.0"
SQLAlchemy = "1.4.45"
starlette = "0.22.0"
typing-extensions = "^4.4.0"
ujson = "5.6.0"
urllib3 = "1.26.14"
uvicorn = "0.20.0"
uvloop = "0.17.0"
watchfiles = "^0.18.1"
wcwidth = "^0.2.5"
websockets = "^10.4"
Werkzeug = "^2.2.2"
wheel = "^0.38.4"
zipp = "^3.11.0"
gunicorn = "^20.1.0"
starlette-context = "^0.3.5"
blinker = "^1.5"
Faker = "^16.8.1"
redis = "^4.5.1"
pinject = "^0.14.1"
fakeredis = "^2.9.0"

[tool.poetry.dev-dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
pytest = "^7.2.0"
pre-commit = "^2.21.0"
coverage = "^7.0.5"
pylint = "^2.15.10"
pytest-mock = "^3.10.0"

[tool.black]
line-length = 89
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(/(
    \.git
  | \.hg
  | \.eggs
  | \.ini
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | .yml
  | .yaml
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
)/
)
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
